task9

+++ 0x01-variables_if_else_while/9-print_comb.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	float n;
+	int n;
 
 	for (n = '0' ; n <= '9' ; n++)
 	{
